+++ src/backend/main/go.backends/store/cassandra/contacts.go
@@ -7,11 +7,17 @@
 package store
 
 import (
+	"errors"
 	. "github.com/CaliOpen/Caliopen/src/backend/defs/go-objects"
+	"github.com/gocassa/gocassa"
 )
 
-func (cb *CassandraBackend) GetContact(user_id, contact_id string) (contact *Contact, err error) {
-	contact = &Contact{}
+func (cb *CassandraBackend) CreateContact(contact *Contact) error {
+	return errors.New("[CassandraBackend] not implemented")
+}
+
+func (cb *CassandraBackend) RetrieveContact(user_id, contact_id string) (contact *Contact, err error) {
+	contact = new(Contact).NewEmpty().(*Contact)
 	m := map[string]interface{}{}
 	q := cb.Session.Query(`SELECT * FROM contact WHERE user_id = ? AND contact_id = ?`, user_id, contact_id)
 	err = q.MapScan(m)
@@ -30,6 +36,22 @@ func (cb *CassandraBackend) GetContact(user_id, contact_id string) (contact *Con
 	}
 
 	return contact, err
+}
+
+func (cb *CassandraBackend) UpdateContact(contact *Contact, fields map[string]interface{}) error {
+	contactT := cb.IKeyspace.Table("contact", &Contact{}, gocassa.Keys{
+		PartitionKeys: []string{"user_id", "contact_id"},
+	}).WithOptions(gocassa.Options{TableName: "contact"}) // need to overwrite default gocassa table naming convention
+
+	err := contactT.
+		Where(gocassa.Eq("user_id", contact.UserId.String()), gocassa.Eq("contact_id", contact.ContactId.String())).
+		Update(fields).
+		Run()
+	return err
+}
+
+func (cb *CassandraBackend) DeleteContact(contact *Contact) error {
+	return errors.New("[CassandraBackend] not implemented")
 }
 
 func (cb *CassandraBackend) LookupContactsByIdentifier(user_id, address string) (contact_ids []string, err error) {