+++ src/backend/main/go.backends/store/cassandra/messages.go
@@ -21,7 +21,7 @@ func (cb *CassandraBackend) CreateMessage(msg *Message) error {
 
 func (cb *CassandraBackend) RetrieveMessage(user_id, msg_id string) (msg *Message, err error) {
 
-	msg = &Message{}
+	msg = new(Message).NewEmpty().(*Message) // correctly initialize nested values
 	m := map[string]interface{}{}
 	q := cb.Session.Query(`SELECT * FROM message WHERE user_id = ? and message_id = ?`, user_id, msg_id)
 	err = q.MapScan(m)