+++ src/backend/main/py.main/caliopen_main/message/parameters/message.py
@@ -12,7 +12,6 @@ from .participant import Participant
 from .attachment import Attachment
 from .external_references import ExternalReferences
 from caliopen_main.pi.parameters import PIParameter
-from caliopen_main.common.parameters.tag import ResourceTag
 from caliopen_main.user.parameters import Identity
 import caliopen_storage.helpers.json as helpers
 
@@ -39,11 +38,11 @@ class NewMessage(Model):
     message_id = UUIDType()
     parent_id = UUIDType()
     participants = ListType(ModelType(Participant), default=lambda: [])
-    privacy_features = DictType(StringType, default=lambda: {})
+    privacy_features = DictType(StringType(), default=lambda: {})
     pi = ModelType(PIParameter)
     raw_msg_id = UUIDType()
     subject = StringType()
-    tags = ListType(ModelType(ResourceTag), default=lambda: [])
+    tags = ListType(StringType(), default=lambda: [])
     type = StringType(choices=MESSAGE_TYPES)
 
     class Options: