+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- API for tags management : user's tags, tags on messages and contacts
+
 ### Fixed
 
 - unable to load more messages on Timeline